# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
      parent_id:
        description: nolint
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.City:
    properties:
      id:
        type: integer
      name:
        description: nolint
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError'
      status:
        type: integer
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError:
    properties:
      error:
        type: string
    type: object
...
  internal_pkg_category_delivery.CategoryListResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Category'
        type: array
      status:
        type: integer
    type: object
  internal_pkg_city_delivery.CityListResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.City'
        type: array
      status:
        type: integer
    type: object
  internal_pkg_file_service_delivery.ResponseURLBody:
    properties:
      urls:
        items:
          type: string
        type: array
    type: object
  internal_pkg_file_service_delivery.ResponseURLs:
    properties:
      body:
        $ref: '#/definitions/internal_pkg_file_service_delivery.ResponseURLBody'
      status:
        type: integer
    type: object
...
host: 84.23.53.28:8080
info:
  contact: {}
  description: This is a server of YULA server.
  title: YULA project API
  version: "1.0"
paths:
  /category/get_full:
    get:
      description: get all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_category_delivery.CategoryListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get all categories
      tags:
      - category
  /city/get_full:
    get:
      description: get all cities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_city_delivery.CityListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get all cities
      tags:
      - City
  /img/:
    get:
      consumes:
      - application/json
      description: download photo of file by its name
      parameters:
      - description: name of image
        in: path
        name: name
        required: true
        type: string
      produces:
      - image/png
      - image/jpeg
      - application/json
      responses:
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: download photo
      tags:
      - fileService
  /img/upload:
    post:
      consumes:
      - multipart/form-data
      description: upload photo to file service and return its url
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_file_service_delivery.ResponseURLs'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: upload photo
      tags:
      - fileService
...
schemes:
- http
swagger: "2.0"